requirements.txt
bcrypt
fastapi
fastapi[standard]
google-generativeai
jinja2
joblib
numpy==1.24.3
passlib
pillow
protobuf==3.19.6
pydantic
python-dotenv
python-jose
python-multipart
requests
scikit-learn==1.5.2
scipy
sqlalchemy
tensorflow==2.10.0
uvicorn